Study Feature
Study feature is running as expected: Study page is routed from the path "/decks/:deckId/study" in the Layout.js file, to run the study button on the deck info screen. When study is clicked, the first card starts from the front side, you are able to hit the flip button to see the other side of the card, and then move on to the next card with the next button.